requirements.txt
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
gspread==6.0.2
tweepy==4.14.0
requests==2.31.0
openai==1.12.0
anthropic==0.18.1
sqlalchemy==2.0.25
apscheduler==3.10.4
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pandas==2.2.0
python-dateutil==2.8.2